HW-2 Small fixes and refactoring Logger class
